Remove almost all the input keys create inputs.py, validations.py and utils.py
