+++ performance-evaluation/constants.py
@@ -31,6 +31,7 @@ TEMPLATE_TALENT_FILE_KEY = '1D04Q-IQ67F1wTAgk3oEr1Q902DJd4ulv666mBlcar6c'
 # For file named: 'Rol Laboral - Evaluaciones de Desempeño (Respuestas)'
 ANSWERS_ROLE_FILE_KEY = '1felT_0RAVlG4FWFTbCkx3XMJjVSTd5sXqOLVYMzcRSo'
 
+OPERATION_CREATE_SHEET = "OPERATION_CREATE_SHEET"
 OPERATION_COPY_TABS = "OPERATION_COPY_TABS"
 OPERATION_BUILD_EVALUATION_FORM = "OPERATION_BUILD_EVALUATION_FORM"
 OPERATION_HIDE_TALENTS = "OPERATION_HIDE_TALENTS"
@@ -45,6 +46,7 @@ PERFORMANCE_STUDY_FOLDER_KEY = "1wGrZ9n-il5YnaS6Z_ihapCg1z81ad5Bn"
 PERFORMANCE_REPORT_FOLDER_KEY = "0B-gOApkHyMPvelYzM3VGV2ZlYVk"
 
 operations_by_mode_dictionary = {
+    OPERATION_CREATE_SHEET:             [AUTO_EVALUATION, MANAGER_EVALUATION, EXCHANGE_EVALUATION, FIRST_EVALUATION],
     OPERATION_COPY_TABS: 				[NEXT_EVALUATION, AUTO_EVALUATION, MANAGER_EVALUATION, EXCHANGE_EVALUATION, FIRST_EVALUATION, RID_EVALUATION],
     OPERATION_BUILD_EVALUATION_FORM:	[EXCHANGE_EVALUATION, FIRST_EVALUATION],
     OPERATION_HIDE_TALENTS: 			[NEXT_EVALUATION, AUTO_EVALUATION, MANAGER_EVALUATION, EXCHANGE_EVALUATION, FIRST_EVALUATION],
