+++ performance-evaluation/utils.py
@@ -1,0 +1,14 @@
+import time
+
+def wait_for_quota_renewal():
+	remaining_seconds = 20
+	waiting_seconds = 2
+	while remaining_seconds > 0:
+		print('Por favor esperar ' + str(remaining_seconds) + ' segundos...')
+		time.sleep(waiting_seconds)
+		remaining_seconds = remaining_seconds - waiting_seconds
+	print('Continuando ejecución...')
+	print('')
+
+def on_finish():
+	print('Ejecución finalizada con éxito! No olvidar verificar el estado del documento de forma manual.')
